Implemented GenericCommandBuffer, TODO actually use it
